% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_query_and_bh_refs.R
\name{align_query_and_bh_refs}
\alias{align_query_and_bh_refs}
\title{Title Align query and best hit references}
\usage{
align_query_and_bh_refs(
  query_nr,
  best_hits,
  ref_data,
  marker_data,
  query_list,
  out.path
)
}
\arguments{
\item{query_nr}{Query number (ID)}

\item{best_hits}{List of best hit data}

\item{ref_data}{List of reference data}

\item{marker_data}{List of markers}

\item{query_list}{List of query data}

\item{out.path}{Output path}
}
\value{
Creates table of aligned query and best hit references
}
\description{
Title Align query and best hit references
}
\examples{
align_query_and_bh_refs(query_nr,best_hits,ref_data,marker_data,query_list,out.path)

}
