% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caos_classifier.R
\name{caos_classifier}
\alias{caos_classifier}
\title{Title CAOS-Classifier}
\usage{
caos_classifier()
}
\value{
Classifies taxa based on diagnostic characters
}
\description{
Title CAOS-Classifier
}
\examples{
caos_classifier()
}
