% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Title Turn vector into branch separated list}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{CA_ref}

\item{n}{branch_nr}
}
\value{
Return branch separated list
}
\description{
Title Turn vector into branch separated list
}
\examples{
chunk(CA_ref,branch_nr)
}
