% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_query_with_ref.R
\name{combine_query_with_ref}
\alias{combine_query_with_ref}
\title{Title Combine query with ref}
\usage{
combine_query_with_ref(bh_list, query_list, bh, marker_data, align)
}
\arguments{
\item{bh_list}{best hit list}

\item{query_list}{query list}

\item{bh}{best hit}

\item{marker_data}{marker data}

\item{align}{alignment}
}
\value{
Aligned table
}
\description{
Title Combine query with ref
}
\examples{
combine_query_with_ref(bh_list,query_list,bh,marker_data,align)
}
