% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_barcodes_vs_query.R
\name{compare_barcodes_vs_query}
\alias{compare_barcodes_vs_query}
\title{Title Compare barcodes with query data}
\usage{
compare_barcodes_vs_query(
  node_pos,
  marker_data,
  ref_node,
  input.path,
  out.path,
  query_list,
  marker_priority
)
}
\arguments{
\item{node_pos}{Node position}

\item{marker_data}{Marker data}

\item{ref_node}{Reference node}

\item{input.path}{Input path}

\item{out.path}{Output path}

\item{query_list}{Query list}

\item{marker_priority}{Marker priority}
}
\value{
Save hit score, filtered hit score, score overview and filtered score overview
}
\description{
Title Compare barcodes with query data
}
\examples{
compare_barcodes_vs_query(node_pos,marker_data,ref_node,input.path,out.path,query_list,marker_priority)
}
