% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_patterns.R
\name{control_patterns}
\alias{control_patterns}
\title{Title Control if patterns in data are matched by allowed patterns}
\usage{
control_patterns(marker.CAs, marker_pattern)
}
\arguments{
\item{marker.CAs}{Marker character attributes}

\item{marker_pattern}{Allowed diagnostics}
}
\value{
Return only character attributes that match reference diagnostics (Marker Patterns)
}
\description{
Title Control if patterns in data are matched by allowed patterns
}
\examples{
control_patterns(marker.CAs, marker_pattern)
}
