% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_align_symbols.R
\name{create_align_symbols}
\alias{create_align_symbols}
\title{Title Create alignment symbols ("|"=match , ""=mismatch)}
\usage{
create_align_symbols(bh_list, query_list)
}
\arguments{
\item{bh_list}{Best hit list}

\item{query_list}{Query list}
}
\value{
Return modified alignment
}
\description{
Title Create alignment symbols ("|"=match , ""=mismatch)
}
\examples{
create_align_symbols(bh_list,query_list)
}
