% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_list.R
\name{create_query_list}
\alias{create_query_list}
\title{Title Create query list (1:Query_name,2-n:Markerdata (separated))}
\usage{
create_query_list(query)
}
\arguments{
\item{query}{Query data}
}
\value{
Query list
}
\description{
Title Create query list (1:Query_name,2-n:Markerdata (separated))
}
\examples{
create_query_list(query)
}
