% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_ref_no_score.R
\name{delete_ref_no_score}
\alias{delete_ref_no_score}
\title{Title Delete node data with no score}
\usage{
delete_ref_no_score(hit_score, ref_taxa_data)
}
\arguments{
\item{hit_score}{Hit Score}

\item{ref_taxa_data}{Reference taxa data}
}
\value{
Modified reference taxa data
}
\description{
Title Delete node data with no score
}
\examples{
delete_ref_no_score(hit_score,ref_taxa_data)
}
