% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bh_names.R
\name{extract_bh_names}
\alias{extract_bh_names}
\title{Title Extract best hit (bh) names}
\usage{
extract_bh_names(query_nr, best_hits)
}
\arguments{
\item{query_nr}{Query number (ID)}

\item{best_hits}{Best hits}
}
\value{
Best hit names
}
\description{
Title Extract best hit (bh) names
}
\examples{
extract_bh_names(query_nr,best_hits)
}
