% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diagnostics.R
\name{filter_diagnostics}
\alias{filter_diagnostics}
\title{Title Filter diagnostics}
\usage{
filter_diagnostics(data)
}
\arguments{
\item{data}{Character data}
}
\value{
Return modified character data
}
\description{
Title Filter diagnostics
}
\examples{
filter_diagnostics(CA_subset$All_Characters)
}
