% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_overview.R
\name{filter_overview}
\alias{filter_overview}
\title{Title Filter score for hits}
\usage{
filter_overview(filters, OV_score)
}
\arguments{
\item{filters}{Filter (sorts out nodes without matching diagnostics)}

\item{OV_score}{Hit score}
}
\value{
Return filtered hit score
}
\description{
Title Filter score for hits
}
\examples{
filter_overview(filters, score_overview)
}
