% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_score.R
\name{filter_score}
\alias{filter_score}
\title{Title Filter score for hits}
\usage{
filter_score(filters, hit_score)
}
\arguments{
\item{filters}{Filters (sorts out nodes without matching diagnostics)}

\item{hit_score}{Hit score}
}
\value{
Return filtered hit score
}
\description{
Title Filter score for hits
}
\examples{
filter_score(filters, hit_score)
}
