% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_uni_CAs.R
\name{filter_uni_CAs}
\alias{filter_uni_CAs}
\title{Title Filter unique CAs in array "unique.CAs.characters"}
\usage{
filter_uni_CAs(unique.CAs.characters)
}
\arguments{
\item{unique.CAs.characters}{Unique CA characters}
}
\value{
Return true unique CA characters
}
\description{
Title Filter unique CAs in array "unique.CAs.characters"
}
\examples{
filter_uni_CAs(unique.CAs.characters)
}
