% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_hits.R
\name{find_best_hits}
\alias{find_best_hits}
\title{Title Identify and save best hit reference taxa from nodes with hits}
\usage{
find_best_hits(hit_score, ref_taxa_data, unique_ref_taxa_data)
}
\arguments{
\item{hit_score}{Hit score}

\item{ref_taxa_data}{Reference taxa data}

\item{unique_ref_taxa_data}{Unique reference taxa data}
}
\value{
Return leading taxa
}
\description{
Title Identify and save best hit reference taxa from nodes with hits
}
\examples{
find_best_hits(hit_score,ref_taxa_data,unique_ref_taxa_data)
}
