% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diagnostic_characters.R
\name{find_diagnostic_characters}
\alias{find_diagnostic_characters}
\title{Title Identify diagnostic characters}
\usage{
find_diagnostic_characters(
  tree.nr,
  nodes.tree,
  all.branch.names,
  CAOS,
  empty_row,
  overview.path
)
}
\arguments{
\item{tree.nr}{Tree number}

\item{nodes.tree}{Tree nodes}

\item{all.branch.names}{All branch names}

\item{CAOS}{CAOS sorted list of all processed input data}

\item{empty_row}{Placeholder positions to separate branch data}

\item{overview.path}{Overview path}
}
\value{
Return tables with diagnostic characters
}
\description{
Title Identify diagnostic characters
}
\examples{
find_diagnostic_characters(tree.nr,nodes.tree,all.branch.names,CAOS,empty_row,overview.path)
}
