% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_diagnostics_per_branch.R
\name{find_diagnostics_per_branch}
\alias{find_diagnostics_per_branch}
\title{Title Find unique diagnostics for loaded node}
\usage{
find_diagnostics_per_branch(branch_CA, branches)
}
\arguments{
\item{branch_CA}{Branch character attributes}

\item{branches}{Branch data}
}
\value{
Return list of diagnostics for loaded node
}
\description{
Title Find unique diagnostics for loaded node
}
\examples{
find_diagnostics_per_branch(branch_CA,branches)
}
