% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_match_query_vs_ref.R
\name{find_match_query_vs_ref}
\alias{find_match_query_vs_ref}
\title{Title Find query match with simple pure/private diagnostic}
\usage{
find_match_query_vs_ref(CA_query, ref, match, match_control, points)
}
\arguments{
\item{CA_query}{Character attribute query data}

\item{ref}{Reference data}

\item{match}{Match}

\item{match_control}{Match control}

\item{points}{Points}
}
\value{
Return matches between query and reference data
}
\description{
Title Find query match with simple pure/private diagnostic
}
\examples{
find_match_query_vs_ref(CA_query,ref,match,match_control,points)
}
