% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_match_query_vs_ref2.R
\name{find_match_query_vs_ref2}
\alias{find_match_query_vs_ref2}
\title{Title Find query match with unique diagnostic}
\usage{
find_match_query_vs_ref2(CA_query, ref, match, match_control, points)
}
\arguments{
\item{CA_query}{Character attribute query}

\item{ref}{Reference data}

\item{match}{Match}

\item{match_control}{Match control}

\item{points}{Points}
}
\value{
Return matching data between query and unique diagnostics
}
\description{
Title Find query match with unique diagnostic
}
\examples{
find_match_query_vs_ref2(CA_query,ref,match,match_control,points)
}
