% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barcode_location.R
\name{get_barcode_location}
\alias{get_barcode_location}
\title{Title Load reference barcode location for each marker}
\usage{
get_barcode_location(marker_data, ref_node, node)
}
\arguments{
\item{marker_data}{Marker data}

\item{ref_node}{Reference node}

\item{node}{Node data}
}
\value{
Return barcode location
}
\description{
Title Load reference barcode location for each marker
}
\examples{
get_barcode_location(marker_data,ref_node,node)
}
