% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bh_ref_data.R
\name{get_bh_ref_data}
\alias{get_bh_ref_data}
\title{Title Create list with reference marker data}
\usage{
get_bh_ref_data(bh, ref_data)
}
\arguments{
\item{bh}{Best hit}

\item{ref_data}{Reference data}
}
\value{
Return best hit data
}
\description{
Title Create list with reference marker data
}
\examples{
get_bh_ref_data(bh, ref_data)
}
