% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branch_hierarchy.R
\name{get_branch_hierarchy}
\alias{get_branch_hierarchy}
\title{Title Load branch hierarchy}
\usage{
get_branch_hierarchy(tree)
}
\arguments{
\item{tree}{Tree file}
}
\value{
Return tree hierarchy
}
\description{
Title Load branch hierarchy
}
\examples{
get_branch_hierarchy(tree)
}
