% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CA_per_taxa.R
\name{get_CA_per_taxa}
\alias{get_CA_per_taxa}
\title{Title Save CAs for each branch, taxa and position in list}
\usage{
get_CA_per_taxa(branches, CAOS)
}
\arguments{
\item{branches}{Branches}

\item{CAOS}{CAOS processed input database}
}
\value{
Return branch character attributes
}
\description{
Title Save CAs for each branch, taxa and position in list
}
\examples{
get_CA_per_taxa(branches,CAOS)
}
