% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_data.R
\name{get_marker_data}
\alias{get_marker_data}
\title{Title Split FASTA data into marker related data}
\usage{
get_marker_data(fasta.names, fasta.seq, mark_nr)
}
\arguments{
\item{fasta.names}{Fasta names}

\item{fasta.seq}{Fasta sequences}

\item{mark_nr}{marker number}
}
\value{
Return marker sorted database of sequence names and data
}
\description{
Title Split FASTA data into marker related data
}
\examples{
get_marker_data(fasta.names,fasta.seq,mark_nr)
}
