% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_diagnostics.R
\name{get_marker_diagnostics}
\alias{get_marker_diagnostics}
\title{Title List diagnostics for each marker}
\usage{
get_marker_diagnostics(input.path, barcode_location)
}
\arguments{
\item{input.path}{Input data path}

\item{barcode_location}{Location of barcode data}
}
\value{
Return marker diagnostics
}
\description{
Title List diagnostics for each marker
}
\examples{
get_marker_diagnostics(input.path,barcode_location)
}
