% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_patterns.R
\name{get_marker_patterns}
\alias{get_marker_patterns}
\title{Title Define allowed patterns of characters}
\usage{
get_marker_patterns(marker_patterns, mark_nr)
}
\arguments{
\item{marker_patterns}{Marker patterns}

\item{mark_nr}{Marker id}
}
\value{
Return allowed marker specific diagnostics
}
\description{
Title Define allowed patterns of characters
}
\examples{
get_marker_patterns(marker_patterns,mark_nr)
}
