% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_diagnostics.R
\name{get_pos_diagnostics}
\alias{get_pos_diagnostics}
\title{Title Get diagnostics character attributes (CA) from reference barcode database}
\usage{
get_pos_diagnostics(CA, row, character_type)
}
\arguments{
\item{CA}{Character attributes}

\item{row}{Data position}

\item{character_type}{Diagnostic character type}
}
\value{
Return reference character attributes
}
\description{
Title Get diagnostics character attributes (CA) from reference barcode database
}
\examples{
get_pos_diagnostics(CA,row,character_type)
}
