% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priority_of_markers.R
\name{get_priority_of_markers}
\alias{get_priority_of_markers}
\title{Title Get priority level for each marker}
\usage{
get_priority_of_markers(ref_path, ref_marker)
}
\arguments{
\item{ref_path}{Path to reference data}

\item{ref_marker}{Reference marker data}
}
\value{
Return marker priority
}
\description{
Title Get priority level for each marker
}
\examples{
get_priority_of_markers(ref_path, ref_marker)
}
