% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_names_and_ca.R
\name{get_taxa_names_and_ca}
\alias{get_taxa_names_and_ca}
\title{Title Get Species names in node and all characters}
\usage{
get_taxa_names_and_ca(node, CAOS)
}
\arguments{
\item{node}{Nodes}

\item{CAOS}{CAOS input database}
}
\value{
Return taxa name and character attribute information
}
\description{
Title Get Species names in node and all characters
}
\examples{
get_taxa_names_and_ca(node,CAOS)
}
