% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_ref_taxa.R
\name{get_unique_ref_taxa}
\alias{get_unique_ref_taxa}
\title{Title Get unique taxa data summary}
\usage{
get_unique_ref_taxa(ref_taxa_data)
}
\arguments{
\item{ref_taxa_data}{Reference taxa data}
}
\value{
Return unique taxon data
}
\description{
Title Get unique taxa data summary
}
\examples{
get_unique_ref_taxa(ref_taxa_data)
}
