% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_diagnostics.R
\name{identify_diagnostics}
\alias{identify_diagnostics}
\title{Title Sort CAs into simple pure and simple private character types}
\usage{
identify_diagnostics(all.CAs.characters, exclusive.unique.CAs.characters)
}
\arguments{
\item{all.CAs.characters}{All character attribute characters}

\item{exclusive.unique.CAs.characters}{Exclusive unique character attribute characters}
}
\value{
Return character attribute types(simple pure & simple private)
}
\description{
Title Sort CAs into simple pure and simple private character types
}
\examples{
identify_diagnostics(all.CAs.characters,exclusive.unique.CAs.characters)
}
