% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_unique_diagnostics.R
\name{identify_unique_diagnostics}
\alias{identify_unique_diagnostics}
\title{Title Find unique diagnostics between branches of the selected node}
\usage{
identify_unique_diagnostics(branch_sep.CA, CA.chars.nr)
}
\arguments{
\item{branch_sep.CA}{branch specific character attribute data}

\item{CA.chars.nr}{Character attribute position}
}
\value{
Return diagnostic data sorted by nodes
}
\description{
Title Find unique diagnostics between branches of the selected node
}
\examples{
identify_unique_diagnostics(branch_sep.CA,CA.chars.nr)
}
