% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_input_data.R
\name{index_input_data}
\alias{index_input_data}
\title{Title Index input data}
\usage{
index_input_data(query_path)
}
\arguments{
\item{query_path}{Path to query data}
}
\value{
Return index of input data
}
\description{
Title Index input data
}
\examples{
index_input_data(query_path)
}
