% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_input_data2.R
\name{index_input_data2}
\alias{index_input_data2}
\title{Title Index input data}
\usage{
index_input_data2(input.path)
}
\arguments{
\item{input.path}{Path to input data}
}
\value{
Return index of input data
}
\description{
Title Index input data
}
\examples{
index_input_data2(input.path)
}
