% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_best_hits.R
\name{print_best_hits}
\alias{print_best_hits}
\title{Title Save and print best hits per query}
\usage{
print_best_hits(query_nr, query_data, leading_taxa, best_hits, out.path)
}
\arguments{
\item{query_nr}{Query number (ID)}

\item{query_data}{Query data}

\item{leading_taxa}{Leading taxa}

\item{best_hits}{Best hit data}

\item{out.path}{Path to output}
}
\value{
Best hits per query
}
\description{
Title Save and print best hits per query
}
\examples{
print_best_hits(query_nr,query_data,leading_taxa,best_hits,out.path)
}
