% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fasta_data.R
\name{process_fasta_data}
\alias{process_fasta_data}
\title{Title Process FASTA data}
\usage{
process_fasta_data(fasta.data)
}
\arguments{
\item{fasta.data}{Fasta data}
}
\value{
Return loaded fasta data
}
\description{
Title Process FASTA data
}
\examples{
process_fasta_data(fasta.data)
}
