% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Title Read tree file}
\usage{
read_tree(TREE.Path, Dataset_Names)
}
\arguments{
\item{TREE.Path}{Path to tree file}

\item{Dataset_Names}{Name of datasets}
}
\value{
Return tree data
}
\description{
Title Read tree file
}
\examples{
read_tree(TREE.Path, Dataset_Names)
}
