% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree_osx.R
\name{read_tree_osx}
\alias{read_tree_osx}
\title{Title Read tree file (Linux/OSX)}
\usage{
read_tree_osx(TREE.Path, Dataset_Names)
}
\arguments{
\item{TREE.Path}{Path to tree file}

\item{Dataset_Names}{Name of datasets}
}
\value{
Return tree data
}
\description{
Title Read tree file (Linux/OSX)
}
\examples{
read_tree_osx(TREE.Path, Dataset_Names)
}
