% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree_win.R
\name{read_tree_win}
\alias{read_tree_win}
\title{Title Read tree file (Windows)}
\usage{
read_tree_win(TREE.Path, Dataset_Names)
}
\arguments{
\item{TREE.Path}{Path to tree file}

\item{Dataset_Names}{Name of datasets}
}
\value{
Return tree data
}
\description{
Title Read tree file (Windows)
}
\examples{
read_tree_win(TREE.Path, Dataset_Names)
}
