% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganize_data.R
\name{reorganize_data}
\alias{reorganize_data}
\title{Title Reorganize ref_taxa based on marker hierarchy}
\usage{
reorganize_data(data, data_order)
}
\arguments{
\item{data}{Data}

\item{data_order}{Order of data}
}
\value{
Return reorganized data
}
\description{
Title Reorganize ref_taxa based on marker hierarchy
}
\examples{
reorganize_data(data,data_order)
}
