% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_branch_taxa.R
\name{separate_branch_taxa}
\alias{separate_branch_taxa}
\title{Title Save taxa names in separate vectors saved in list "branches"}
\usage{
separate_branch_taxa(node)
}
\arguments{
\item{node}{Node}
}
\value{
Return branch data
}
\description{
Title Save taxa names in separate vectors saved in list "branches"
}
\examples{
separate_branch_taxa(node)
}
