% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_characters.R
\name{split_characters}
\alias{split_characters}
\title{Title Separate single/multi-character patterns}
\usage{
split_characters(marker_seqs)
}
\arguments{
\item{marker_seqs}{Marker sequence}
}
\value{
Return character attributes for each marker
}
\description{
Title Separate single/multi-character patterns
}
\examples{
split_characters(marker_seqs)
}
