% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_diagnostics.R
\name{split_diagnostics}
\alias{split_diagnostics}
\title{Title Split diagnostics}
\usage{
split_diagnostics(bh_list)
}
\arguments{
\item{bh_list}{List of best hits}
}
\value{
Return edited list of best hits
}
\description{
Title Split diagnostics
}
\examples{
split_diagnostics(bh_list)
}
